+++ doc/source/CHANGELOG.rst
@@ -15,6 +15,7 @@ Changelog
 - msm: fix connected sets for Bayesian MSMs in case the mincount connectivity (ergodic cutoff) parameter truncated
   the count matrix. #1343
 - plots: fix vmin, vmax keyword arguments for plot_contour(). #1376
+- coordinates: forcefully enable checking of coordinates data streaming for invalid (not initialized) data. #1384